REVERTME: STPK-1171 Improve Camera on Dell XPS12
This commit makes the camera on the Dell XPS12 semi-functional.  It
does this by modifying the kernel uvc driver to understand the
payloads that are received from the Dell XPS 12 camera.  Note that this
is not a perfect fix and there is still some image corruption, although
much less than before.  The issue is either caused by a problem with
the camera itself or some problems with the lower levels of the USB
stack.  If we identify the correct fix we will revert this change.

Finally, this change has been implemented as a quirk which means
that it should only effect the Dell XPS12 camera.  All other cameras
should follow the old code path.

Issue: STPK-1171
Change-Id: If8d58750fad303f3c18b3933eece9b4d1a91de4a
Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>