ACPI battery: fix hard shutdown issue and not showing fully charged
On Broadwell platforms 6% battery is the minimum threshold. When
the reported capacity from EC reaches less than 6% there should
be a hard shutdown. This 6% minimum should be mapped to 0% in android
which is not hapenning in ACPI battery driver. This patch fixes this
after which Android will gracefully shutdown when battery reaches 0%.

Additionally, the charging is remapped when close to fully charged.
For batteries not recently calibrated, the update allows a battery
charged at to be shown as charged when the remapped level reaches
100%.

Based on commit: I32a27d87a265e3f478b7a6798aca561029f3edd4

Issue: GMIN-1579
Change-Id: I8c3ac316cfd0054e49467a466564cc851f114c03
Signed-off-by: Hakan Englund <hakan.englund@intel.com>