atomisp_gmin_platform: Registration API cleanup
The MIPI CSI port number is a schematic wiring detail that was being
queried in the sensor driver senselessly.  Conversely the CSI data
format and Bayer order were being stored in configuration when they
are clearly part of the sensor silicon behavior and belong in the
driver.

Adjust the gmin_camera_platform_data() and
atomisp_register_i2c_module() API to put values where they belong
(which is a little subtle, as some values are needed before
registration to probe the device) and update the 6 gminified sensors
appropriately.

Similarly the i2c_client can be gotten from the v4l2_sudev, so remove
that from the atomisp_register_i2c_module() arguments.

Issue: GMIN-2474
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>