FOR_UPSTREAM [VPG]: drm/i915/chv: Total EU count via GETPARAM
Starting with Cherryview, devices may have a varying number of EU for
a given ID due to creative fusing. The surest way to determine the
configuration is by reading fuses which is best done in the kernel and
communicated to userspace. The immediate need from userspace is to
determine the number of threads of compute work that can be safely
submitted.

Issue: GMIN-3332
Change-Id: I21a313236445ab89202d7b4f00bf938195623214
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>
Signed-off-by: Deepak S <deepak.s@intel.com>