FROM_UPSTREAM [VPG]: drm/i915: Fix unsafe vma iteration in i915_drop_caches
When unbinding, there is a possibility that we drop the active reference
on the object, thereby freeing it. If that happens, we may destroy the
vm link as well as the object and vma. So iterate carefully.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
(cherry picked from commit 4ad72b7fadd285f849439cdbc408f8b847cef704)

Change-Id: Ife64046385a99543503027d6dd561bd49d52d412
Upstream-repo: git://anongit.freedesktop.org/drm-intel
Upstream-commit: 4ad72b7fadd285f849439cdbc408f8b847cef704
Upstream-as-of-tag: drm-intel-next-2014-09-05
For: GMINL-2586
For: VIZ-2020
Signed-off-by: Dave Gordon <david.s.gordon@intel.com>