SQUASHME! [VPG]: drm/i915: New memtracker sysfs interface design for collecting object stats
SQUASHME! - This patch should be squashed into the following existing patch:
Author: Sourab Gupta <sourab.gupta@intel.com>
Date:   Mon Aug 11 13:12:04 2014 +0530
Change-Id: I02552d54a753fcfb94b5bdd9bc9c82702bbf514d
	MUST_REBASE [VPG]: drm/i915: Sysfs interface to get GFX shmem usage
	stats per process

This patch introduces new design for exposing the object stats via sysfs
interface.

Previously, we used to have a single sysfs file for exposing the cumulative
data of all the processes which have gfx buffers allocated.
This leads to overhead for userspace in parsing this file everytime
(even for non-gfx processes), even to find out whether that process is the
consumer of GFX resources or not.

In this patch, we've introduced a new design whereby there will be a
seperate sysfs file created per each process (which has gfx resources
allocated).
This file can then be queried to retrieve the buffer stats for that process.

v2: Protect the read function of sysfs file using mutex, to protect against
    concurrent access by two processes accessing the same obj state buf.

v3: Changes for
    * Handling the case of multiple concurrent reads for same sysfs tgid file
       w.r.t different drm open file instances (of same tgid).
    * Synchronizing the sysfs file remove function with corrensponding read
       functions.

v4: Changes for
    * Checking for tgid ptr itself instead of extracting pid num for comparison.
    * Resetting the obj state buf parameters when a fresh read call lands up
      on an already populated obj state buffer.

Issue: GMINL-3834
Change-Id: I7900401595bf96fd88a9d6f2a3e762b5780d1481
Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>
Signed-off-by: Akash Goel <akash.goel@intel.com>