seccomp: really fix syscall numbers for x86 and x86_64
This fixes a build break on x86/x86_64 introduced by f2caf97
"seccomp: fix syscall numbers for x86 and x86_64".

The original change looks like it was cherry-picked from a kernel
with sys_renameat2, but since this kernel (android-3.14) lacks an
implementation of that syscall, it is not correct to refer to it
in the x86 syscall tables, or vmlinux will fail to link.

Do what ARM does and define the syscall slot but have it wired
up to the sys_ni_syscall / compat_ni_syscall so we retain syscall
binary compatibility with newer kernels.

Change-Id: I3fd1bc15557f0df3cf744bc8279b0985d48b93bf
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>