FOR_UPSTREAM [VPG]: drm/i915: Fix for edp hpd loop
HPD is enabled for all displays by default, due to which eDP gets
hpd calls for VDD on/PPS on and VDD off/PPS off.

But i915 code to handle HPD for eDP turns on VDD for DPCD read and
then turns it off back again. so when the panel is put to sleep and
the edp panel is powered down it generates first HPD. The code to
handle this hpd turns on HPD and turns if off back again, which again
generates another HPD. thus becoming an infinite loop.

This patch disables HPD bits for eDP panel alone and this avoid the
scenario in the first place.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-16148
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6962
Change-Id: Ia619cb7b7c31724fea9ea3061ab29ecb735a340d
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>