WA: power: power_supply_charger: No C7 during charger handling
Some PMICs cannot deal with rapid transitions into/out of S0i3
effectively. Internal blocks on IC may shutdown during S0i3 and
it takes time to reactivate them, for example, ADC.
This workardound in the charger framework prevents the system
from going into S0i3 during interrogation of PMIC for various
status updates.
This keeps the necessary blocks on PMIC active during this period,
preventing phantom bits from being read, and erroneously writing
incorrect bits into registers.

For devices that do not have external pull-up resistors on
the I2C bus, the internal pull-up may take some time to take
effect. This is further complicated by the fact that the kernel
does not have direct control of suspend/resume on the I2C
controller with PMIC. That is taken care of by PUNIT.

So as a workaround, keep the system out of deep C-state
such that the PMIC is all active, and I2C controller does not
go into suspend. This is enabled by a quirk that is specifically
requested by the PMIC driver.

Change-Id: I9fdf76243d950353a5c40331d66822ab67b8c4e4
Signed-off-by: Daniel Leung <daniel.leung@linux.intel.com>