gt9xx: Change maximum number of supported touches
This driver was written based on GT911 datasheet, which supported
only 5 touches.
We now have GT9271 and GT928, which can detect a maximum of 10 touches.

The extra point received from the firmware wasn't handled correctly in
userspace when the maximum number was reached. This happened because the
driver didn't reported the extra touch, but it did sent the release of
the point when they were release progressively.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-8869
Change-Id: Ibe0914eda2747c4e0feae51fad1317d944ebd38b
Signed-off-by: Irina Patru <irina.patru@intel.com>