REVERTME [VPG]: drm/i915: Trigger flip on all active planes for z-order change
Currently when the z-order is changed flip is not forced on all active
planes and this manifest as flickers in some layers. To avoid this,
when ever the z-order changes, force flip on all active planes by
writing to the trigger register.

REVERTME: This patch fixes an issue in VPG internal set_display IOCTL.
This entire IOCTL and related fixes can be reverted once the upstream
nuclear page flip / atomic mode set lands in this product kernel.

Issue: IMINAN-35785
Change-Id: Iad721e40727a8fdd15b5711509cf96d39f8351d0
Signed-off-by: Arun R Murthy <arun.r.murthy@intel.com>