REVERTME [VPG]: drm/i915: Avoid vdd on/off in detect
Disabling VDD through delayed queue results in timing issue
that degrades the performance of device. This patch avoids
VDD modification in intel_dp_detect as part of OUI read.
It is removed totally as OUI is not needed for proper
functioning of panel.

REVERTME because this issue needs further debugging to find
proper root cause. allowing this patch based on debug done by
Li Ting.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-16148
Change-Id: Iac302baa970e74877993b44103bdb2e60552ea31
Signed-off-by: Sivakumar Thulasimani <sivakumar.thulasimani@intel.com>