ASoc:Intel:BYT: Emdoor board configuration updates
Due to electrical handling among devices, the GPIO used
in jack detection (JACK_DET_N) is not stable enough. As
a result, delays or not change state are found the GPIO
during HP and HS removal. In order to avoid the problem,
audio chip GPIO is used (AUDIO_INT) in jack detection.

Even though GPIO's are in the VLV2 chip, the GPIO states
depends on the audio codec chip (AUDIO_INT) and the VLV2
itself (JACK_DET_N). Same board configuration will be
used for all Emdoor i8xxx variants.

Change-Id: I9561e7daaa3296f712b761cbcc9b402720cc0b51
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-15389
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>